Moved functionality out to new duplicate checker class
